--- app/pages/index.js
@@ -1,3 +1,0 @@
-import UserProfilePage from './UserProfilePage/UserProfilePage';
-
-export { UserProfilePage };
